requirements.txt
Flask==3.0.3
tensorflow==2.20.0
numpy==2.1.1
Pillow==10.4.0
gunicorn==23.0.0
flask-limiter==3.9.2
Werkzeug==3.0.4
python-multipart==0.0.9
